componentへの切り出し
